test(geo): isolate prompt answer hook module mocks
